Added a forms folder
I have added a forms folder that contains a form project. In this project, I have included all the HTML form methods available. I also made it beautiful using CSS.
Anyone who wants to learn HTML forms can learn from it. Hope it will be helpful.
The output ScreenShots are also included in the folder.
Please merge it if it is valid. THANKS!!